pegdownという文字列を除去。
